trying to make the view model work specifically so that there will always be a user's details returned, specifically the current, signed-in user
