+++ frontend/src/main.tsx
@@ -4,7 +4,7 @@ import App from './App'
 import { AuthProvider } from './context/AuthContext'
 import './amplify-config'
 import './styles/global.css'
-import Profile from './pages/Profile'
+import Profile from './Views/Profile'
 
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
